refactor: make read_batch object safe in WalManager (#119)
* refactor: make read_batch object safe in WalManager

* Make TableLogIterator as main branch

* feat: return Vec<u8> in BlockingLogIterator

* fix: impl debug for MetaUpdateReaderImpl

* fix: remove debug print

* feat: change item of BlockingLogIterator &[u8]

* fix: resolve comment
