+++ wal/src/table_kv_impl/region.rs
@@ -23,9 +23,9 @@ use table_kv::{
 use tokio::sync::Mutex;
 
 use crate::{
-    log_batch::{LogEntry, LogWriteBatch, Payload, PayloadDecoder},
+    kv_encoder::LogKey,
+    log_batch::{LogEntry, LogWriteBatch, Payload},
     manager::{self, BlockingLogIterator, ReadContext, ReadRequest, RegionId, SequenceNumber},
-    rocks_impl::encoding::LogKey,
     table_kv_impl::{
         encoding, encoding::LogEncoding, model::RegionEntry, namespace::BucketRef, WalRuntimes,
     },
@@ -52,9 +52,7 @@ pub enum Error {
     },
 
     #[snafu(display("Failed to do log codec, err:{}.", source))]
-    LogCodec {
-        source: crate::rocks_impl::encoding::Error,
-    },
+    LogCodec { source: crate::kv_encoder::Error },
 
     #[snafu(display("Failed to scan table, err:{}", source))]
     Scan {
@@ -546,6 +544,7 @@ impl Region {
 
 pub type RegionRef = Arc<Region>;
 
+#[derive(Debug)]
 pub struct TableLogIterator<T: TableKv> {
     buckets: Vec<BucketRef>,
     /// Inclusive max log key.
@@ -559,6 +558,8 @@ pub struct TableLogIterator<T: TableKv> {
     // The `current_iter` should be either a valid iterator or None.
     current_iter: Option<T::ScanIter>,
     log_encoding: LogEncoding,
+    // TODO(ygf11): Remove this after issue#120 is resolved.
+    previous_value: Vec<u8>,
 }
 
 impl<T: TableKv> TableLogIterator<T> {
@@ -572,6 +573,7 @@ impl<T: TableKv> TableLogIterator<T> {
             current_bucket_index: 0,
             current_iter: None,
             log_encoding: LogEncoding::newest(),
+            previous_value: Vec::default(),
         }
     }
 
@@ -591,6 +593,7 @@ impl<T: TableKv> TableLogIterator<T> {
             current_bucket_index: 0,
             current_iter: None,
             log_encoding: LogEncoding::newest(),
+            previous_value: Vec::default(),
         }
     }
 
@@ -661,10 +664,7 @@ impl<T: TableKv> TableLogIterator<T> {
 }
 
 impl<T: TableKv> BlockingLogIterator for TableLogIterator<T> {
-    fn next_log_entry<D: PayloadDecoder>(
-        &mut self,
-        decoder: &D,
-    ) -> manager::Result<Option<LogEntry<D::Target>>> {
+    fn next_log_entry(&mut self) -> manager::Result<Option<LogEntry<&'_ [u8]>>> {
         if self.no_more_data() {
             return Ok(None);
         }
@@ -691,19 +691,24 @@ impl<T: TableKv> BlockingLogIterator for TableLogIterator<T> {
             .context(manager::Decoding)?;
         let payload = self
             .log_encoding
-            .decode_value(current_iter.value(), decoder)
+            .decode_value(current_iter.value())
             .map_err(|e| Box::new(e) as _)
             .context(manager::Encoding)?;
-        let log_entry = LogEntry {
-            sequence: self.current_log_key.1,
-            payload,
-        };
+
+        // To unblock pr#119, we use the following to simple resolve borrow-check error.
+        // detail info: https://github.com/CeresDB/ceresdb/issues/120
+        self.previous_value = payload.to_owned();
 
         // Step current iterator, if it becomes invalid, reset `current_iter` to None
         // and advance `current_bucket_index`.
         self.step_current_iter()
             .map_err(|e| Box::new(e) as _)
             .context(manager::Read)?;
+
+        let log_entry = LogEntry {
+            sequence: self.current_log_key.1,
+            payload: self.previous_value.as_slice(),
+        };
 
         Ok(Some(log_entry))
     }
